細かな修正

+++ resources/views/applications/show.blade.php
@@ -13,13 +13,15 @@
             <p><a href={{ $app->app_url }}>{{ $app->app_url }}<a/></p>
         </div>
         <div class="footer">
-        <div class="delete">
-            <form method="POST" action="/apps/{{ $app->id }}" id="form_{{ $app->id }}">
-                @csrf
-                @method("DELETE")
-                <input type="button" onclick="deleteProject({{ $app->id }})" value="投稿を削除" />
-            </form>
-        </div>
+            <div class="delete">
+                @if(Auth::user()->id === $app->project->user_id)
+                    <form method="POST" action="/apps/{{ $app->id }}" id="form_{{ $app->id }}">
+                        @csrf
+                        @method("DELETE")
+                        <input type="button" onclick="deleteProject({{ $app->id }})" value="投稿を削除" />
+                    </form>
+                @endif
+            </div>
             <a href="{{ route('apps.list') }}">成果物一覧に戻る</a>
         </div>
         <script>
